+++ src/app/page.tsx
@@ -1,25 +1,12 @@
-import BooksPanel from "@/components/books-panel/books-panel";
+import BooksPanel from "@/components/books/books-panel/books-panel";
 import { BookModel } from "@/models/models";
 import { getAllBooks } from "@/services/lib.service";
 import { Suspense  } from "react";
 import type { Metadata } from 'next'
 import { METADATA_INFOS } from "@/utils/constants";
+import { calculateMetadata } from "@/utils/calculateMetadata";
 
-export const metadata: Metadata = {
-  title: METADATA_INFOS.TITLE,
-  description: METADATA_INFOS.DESCRIPTION,
-  keywords: ['auto-édition', 'livres auto-édités', 'librairie pour auto-édités'],
-  metadataBase: new URL('https://la-librairie-des-autoedites.ducafeetdesrimes.com'),
-  openGraph: {
-    images: ['/images/og.webp']
-  },
-  twitter: {
-    card: 'summary_large_image',
-    images: ['/images/og.webp'],
-    description: METADATA_INFOS.DESCRIPTION,
-    title: METADATA_INFOS.TITLE
-  },
-}
+export const metadata: Metadata = calculateMetadata(METADATA_INFOS.TITLE, '/','images/og.webp')
 
 export default async function Home() {