+++ src/components/books/book-card/book-card.tsx
@@ -1,9 +1,21 @@
+'use client'
 import { BookModel } from "@/models/models";
+import { sendGAEvent } from "@next/third-parties/google";
+import { useRouter } from "next/navigation";
 import { TbExternalLink } from "react-icons/tb";
 
 function BookCard(props: BookModel) {
+
+  const router = useRouter();
+
+  function handleClickOnBuyBook(event:any, link:string){
+    event.preventDefault();
+    sendGAEvent('event', 'buttonClicked', { value: props.title })
+    window.open(link,'_blank');
+  }
+
   return (
-    <div className="card bg-base-100  shadow-xl ">
+    <div className="card bg-white shadow-xl ">
       <figure className="w-full ">
         <img className="object-contain h-80" src={props.cover} alt={'Couverture de ' + props.title} />
       </figure>
@@ -22,7 +34,7 @@ function BookCard(props: BookModel) {
         <a target="_blank" href={props.social_link} className="btn">
               Découvrir l&apos;auteurice <TbExternalLink />
             </a>
-            <a target="_blank" href={props.link} className="btn btn-primary">
+            <a target="_blank" onClick={e => handleClickOnBuyBook(e, props.link)} href={props.link} className="btn btn-primary text-neutral">
               Acheter <TbExternalLink />
             </a>
         </div>