+++ CHANGELOG.md
@@ -1,7 +1,57 @@
-[Unreleased]
+# Changelog
 
-[1.0.0]
+All notable changes to this project will be documented in this file.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [2.0.0] - 15-12-2024
+
+### Fixed
+
+- [#25] 25-fix-have-better-slugs
+- [#10] 10-fix-navbar-blog-outline-in-blog-articles-part
+
+### Added 
+
+- [#29] 29-add-portail-de-lae-page
+- [#21] 21-add-a-propos-page
+- [#20] 20-generate-metadata
+- [#16] 16-add-sitemap
+- [#9] 9-add-google-analytics
+- [#6] 6-create-blog-section 
+
+### Changed 
+
+- [#31] 31-improve-blog-presentation
+- [#30] 30-improve-a-propos-page
+- [#5] 5-improve-hero-aesthetic : Improve general aesthetic
+- [#4] 4-add-librairie-insta-in-social-links 
+- [#2] feat/evol-navigation : put book category navigation in the panel
+
+## [1.0.0] - 09-10-2024
 
 - Navbar
 - Panel with book cards
 - Search bar
+
+
+<!-- ISSUES -->
+[#2]: https://github.com/willfynch/lalibrairiedesautoedites/issues/2
+[#4]: https://github.com/willfynch/lalibrairiedesautoedites/issues/4
+[#5]: https://github.com/willfynch/lalibrairiedesautoedites/issues/5
+[#6]: https://github.com/willfynch/lalibrairiedesautoedites/issues/6
+[#9]: https://github.com/willfynch/lalibrairiedesautoedites/issues/9
+[#10]: https://github.com/willfynch/lalibrairiedesautoedites/issues/10
+[#16]: https://github.com/willfynch/lalibrairiedesautoedites/issues/16
+[#20]: https://github.com/willfynch/lalibrairiedesautoedites/issues/20
+[#21]: https://github.com/willfynch/lalibrairiedesautoedites/issues/21
+[#25]: https://github.com/willfynch/lalibrairiedesautoedites/issues/25
+[#29]: https://github.com/willfynch/lalibrairiedesautoedites/issues/29
+[#30]: https://github.com/willfynch/lalibrairiedesautoedites/issues/30
+[#31]: https://github.com/willfynch/lalibrairiedesautoedites/issues/31
+
+
+<!-- TAGS -->
+[1.0.0]: https://github.com/willfynch/lalibrairiedesautoedites/releases/tag/v1.0.0
+[2.0.0]: https://github.com/willfynch/lalibrairiedesautoedites/releases/tag/v2.0.0